extract emoji handling logic into its own hook
